requirements.txt
paho-mqtt>=2.0,<3
flask>=3.0,<4
pyyaml>=6.0,<7
